+++ packages/type-eval/src/print/json.ts
@@ -1,10 +1,11 @@
 import { assertExists, assertUnreachable } from "@noom/symbolism-utils";
 import { JsonObject } from "type-fest";
-import ts from "typescript";
 import { isNumericSchema } from "../classify";
 import type { AnySchemaNode, Schema } from "../schema";
 import { schemaToRegEx } from "../string";
 import { printSchemaNode } from "./typescript";
+import { SymbolDisplayPart } from "typescript";
+import { getDocumentationComment } from "@noom/symbolism-ts-utils";
 
 export function createJsonSchema(params: {
   schema: Schema;
@@ -29,25 +30,49 @@ export function createJsonSchema(params: {
   };
 }
 
+const getDocumentation = (schema: AnySchemaNode | undefined) => {
+  const documentationComment = getDocumentationComment(schema?.node)
+    // Split on line break
+    .map((c: SymbolDisplayPart) => c?.text.split(/\r?\n/))
+    .flat(Infinity);
+
+  if (documentationComment.length > 0) {
+    return { documentationComment };
+  }
+
+  return {};
+};
+
 export function schemaToJson(
   schema: AnySchemaNode | undefined
 ): JsonObject | null {
   if (!schema) {
     return null;
   }
+  const documentation = getDocumentation(schema);
 
   switch (schema.kind) {
     case "primitive":
-      return { type: schema.name };
+      return {
+        type: schema.name,
+        ...documentation,
+      };
     case "literal":
       if (typeof schema.value === "bigint") {
-        return { const: `"${schema.value}"` };
+        return {
+          const: `"${schema.value}"`,
+          ...documentation,
+        };
       }
-      return { const: schema.value ?? null };
+      return {
+        const: schema.value ?? null,
+        ...documentation,
+      };
     case "array":
       return {
         type: "array",
         items: schemaToJson(schema.items),
+        ...documentation,
       };
     case "tuple":
       return {
@@ -55,13 +80,11 @@ export function schemaToJson(
         // TODO: Variable length tuples
         maxItems: schema.items.length,
         minItems: schema.items.length,
-        prefixItems: schema.items.map((item, i) => {
+        prefixItems: schema.items.map((item) => {
           // TODO: Emit flags
-          const elementFlags = schema.elementFlags[i];
-          const isRest = elementFlags & ts.ElementFlags.Rest;
-          const isOptional = elementFlags & ts.ElementFlags.Optional;
           return schemaToJson(item);
         }),
+        ...documentation,
       };
     case "object":
       return {
@@ -78,11 +101,13 @@ export function schemaToJson(
               return acc;
             }, {} as Record<string, JsonObject | null>)
           : undefined,
+        ...documentation,
       };
     case "function":
       return {
         type: "error",
         message: `${printSchemaNode(schema)} is not supported in JSON schema`,
+        ...documentation,
       };
     case "binary-expression":
       if (
@@ -91,25 +116,30 @@ export function schemaToJson(
       ) {
         return {
           type: "number",
+          ...documentation,
         };
       }
       return {
         type: "string",
+        ...documentation,
       };
     case "index":
       return {
         type: "error",
         message: `${printSchemaNode(schema)} is not supported in JSON schema`,
+        ...documentation,
       };
     case "index-access":
       return {
         type: "error",
         message: `${printSchemaNode(schema)} is not supported in JSON schema`,
+        ...documentation,
       };
     case "error":
       return {
         type: "error",
         message: JSON.stringify("error! " + schema.extra),
+        ...documentation,
       };
     case "union": {
       const anyOf = schema.items.map(schemaToJson);
@@ -118,24 +148,29 @@ export function schemaToJson(
         return {
           type: "string",
           enum: literals.map((item) => assertExists(item?.const)),
+          ...documentation,
         };
       }
       return {
         anyOf,
+        ...documentation,
       };
     }
     case "intersection":
       return {
         oneOf: schema.items.map(schemaToJson),
+        ...documentation,
       };
     case "template-literal":
       return {
         type: "string",
         pattern: schemaToRegEx(schema) + "",
+        ...documentation,
       };
     case "reference":
       return {
         $ref: `#/$defs/${schema.name}`,
+        ...documentation,
       };
 
     default: