+++ packages/definitions/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @symbolism/definitions
 
+## 0.3.2
+
+### Patch Changes
+
+- 71cbb41: Publish packages
+- Updated dependencies [71cbb41]
+  - @noom/symbolism-ts-debug@0.3.2
+  - @noom/symbolism-ts-utils@0.3.2
+  - @noom/symbolism-utils@0.3.2
+
+## 0.3.1
+
+### Patch Changes
+
+- 9c948b8: Update packages
+- f1283bb: Publish packages
+- Updated dependencies [9c948b8]
+- Updated dependencies [f1283bb]
+  - @noom/symbolism-ts-debug@0.3.1
+  - @noom/symbolism-ts-utils@0.3.1
+  - @noom/symbolism-utils@0.3.1
+
 ## 0.3.0
 
 ### Minor Changes