+++ packages/eslint-plugin/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @symbolism/eslint-plugin
 
+## 0.1.2
+
+### Patch Changes
+
+- 71cbb41: Publish packages
+- Updated dependencies [71cbb41]
+  - @noom/symbolism-definitions@0.3.2
+  - @noom/symbolism-paths@0.3.2
+
+## 0.1.1
+
+### Patch Changes
+
+- 9c948b8: Update packages
+- f1283bb: Publish packages
+- Updated dependencies [9c948b8]
+- Updated dependencies [f1283bb]
+  - @noom/symbolism-definitions@0.3.1
+  - @noom/symbolism-paths@0.3.1
+
 ## 0.1.0
 
 ### Minor Changes